requirements.txt
#
####### Planeshift python requirements #######
# 
# Run pip install -r requirements.txt to install
#
python-daemon
setproctitle

flask
flask_wtf
wtforms-json
flask_sqlalchemy
flask_login

sqlalchemy-migrate
